Remove debugging flushes. Clarify reset behavior.
If a new line is created after there are reset-length lines, then the
buffer is reset.

For example, if there are two lines and the third line updates one of
the first two, a reset will not occur. But if the third line is a new
line, a reset occurs.